rename files to match its default export
